+++ core/admin/panel/Logs.php
@@ -11,11 +11,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 class Logs extends Panel {
-	protected $table = true;
-	protected $sidebar = false;
-	protected $form = true;
-	protected $form_multiform = false;
-	protected $form_method = 'get';
+	protected bool $table = true;
+	protected bool $sidebar = false;
+	protected bool $form = true;
+	protected bool $form_multiform = false;
+	protected string $form_method = 'get';
 
 	public function screen_options_show() {
 		add_screen_option( 'per_page', array(