+++ core/basic/Information.php
@@ -11,11 +11,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 class Information extends BaseInformation {
 	public $code = 'coreactivity';
 
-	public $version = '2.5';
-	public $build = 2500;
+	public $version = '2.6';
+	public $build = 2600;
 	public $edition = 'free';
 	public $status = 'stable';
-	public $updated = '2024.10.15';
+	public $updated = '2024.10.31';
 	public $released = '2023.09.06';
 
 	public $github_url = 'https://github.com/dev4press/coreactivity';