+++ core/build.php
@@ -1,1 +1,1 @@
-<?php $build = 2500;
+<?php $build = 2600;