+++ core/components/Term.php
@@ -179,15 +179,17 @@ class Term extends Component {
 			return;
 		}
 
-		$this->log( 'meta-updated', array(
-			'object_type' => 'term-meta',
-			'object_name' => $meta_key,
-		), array(
-			'user_id'        => $object_id,
+		$args = array(
+			'term_id'        => $object_id,
 			'meta_key'       => $meta_key,
 			'meta_value'     => $meta_value,
 			'meta_value_old' => $old,
-		) );
+		);
+
+		$this->log( 'meta-updated', array(
+			'object_type' => 'term-meta',
+			'object_name' => $meta_key,
+		), $this->validate_old_new( $args, 'meta_value', 'meta_value_old' ) );
 	}
 
 	public function event_meta_deleted( $object_id, $meta_key, $meta_value ) {